update: console suppression mechanism in unit tests
